Adding interactivity to state buffer.
g refreshs the buffer
d marks the process for stoping
m marks the process for muting
u marks teh process for unmuting
x commits the changes

The updates are optimistic and automatically change the buffer and a request is scheduled
to get the actual changes. My next commit(s) will propagate this behaviour to the rest of
the code.