Split emux.el in 3 files. Add emux--dev.el. Add a local structure to keep the known state. Use plists instead of alists.
emux.el is the main file.
emux--plumbing.el contains definitions to support the definition of response and message types and also to deal with json and the process filter.
emux--buffers.el contains definitions to handle what's necessary for *emux*, *emux-log* and *emux-state* buffers to work properly.

Loading emux--dev.el adds its current directory to emacs load-path and defines the function emux--dev-reload-everything that calls load on the previous 3 files.
That is useful when developing. Nevertheless, (require 'emux) works after the directory is in load-path.

I'm keeping a structure of processes that I know are running with as much information as I have about them.
I optimistically update it with the information from messages that are sent/received.
Almost all updates to this running structure are followed by scheduling a state message to get the actual state.
If multiple messages are scheduled, only the last one to be scheduled will be sent.